Update ps_scan.pl: pfsearchV3 cutoff option
Fix: latest pfsearchV3 doesn't recognize -c option, it results in "Option ? is unknown" and aborts the pfsearchV3 run.
With earlier pfsearchV3, the -c option doesn't seem to have the expected effect, i.e. if ps_scan.pl is run with option "-l -1", the cut-off corresponding to the level -1 is put as argument to the -c but only matches with a score higher that the level 0 are retrieved.
Replacing the "-c$cutoff" by "-L $level_min" is valid for the latest pfsearchV3 and has the expected behavior regarding scores/levels for both the latest pfsearchV3 and earlier ones.